Instead of .yaml, allow <dir-name>.yaml
The tool combines both directory and file name to create a SSM path.
When restricting a path with -p, only the corresponding local directory
will be considered; the option to place keys in a directory.yaml file is
not available.

To create SSM entities in that directory, the only way previously
supported was to use a directory/.yaml file (no filename), resulting in SSM
path directory//keys.

This change suppresses the filename if it is the same as the directory
name, allowing directory.yaml to be used (within the directory) instead
of the hidden .yaml file to create keys in the directory/ SSM path.